minor bug fixes in live trade alerts

+++ api/symbol_budget_route_management/dtos/add_budget_setting_dto.py
@@ -10,5 +10,5 @@ class AddBudgetSettingDTO(BaseModel):
     time_frame: Optional[str]
     trades: Trades
     budget_utilization: int
-    start_range: int
-    end_range: int
+    start_range: float
+    end_range: float
