+++ api/symbol_budget_route_management/dtos/modify_budget_setting_dto.py
@@ -11,5 +11,5 @@ class ModifyBudgetSettingDTO(BaseModel):
     time_frame: Optional[str]
     trades: Trades
     budget_utilization: int
-    start_range: int
-    end_range: int
+    start_range: float
+    end_range: float
